Add files via upload
LAN chat application is one of easiest way to chat with a your friends through LAN. No internet connection is needed. The only thing which requires is server IP address and you will be able to connect to others members through LAN . It can help you to talk to your friends even you both do not have internet connection. As it is based on LAN. LAN i.e. local area network which connect different client to each other and also client to main server. So we have used the same concept here we are connecting two client or client and server with each other and by providing the IP address we can talk with each other.